+++ packages/react-query/src/__tests__/ssr.test.tsx
@@ -8,6 +8,8 @@ import {
   QueryClientProvider,
   useInfiniteQuery,
   useIsFetching,
+  useMutation,
+  useMutationState,
   useQueries,
   useQuery,
 } from '..'
@@ -245,6 +247,44 @@ describe('Server Side Rendering', () => {
     expect(markup).toContain('status2: success')
     expect(markup).toContain('data1: data1')
     expect(markup).toContain('data2: data2')
+
+    queryCache.clear()
+  })
+
+  it('useMutation should return idle status', () => {
+    function Page() {
+      const mutation = useMutation({
+        mutationFn: () => sleep(10).then(() => 'data'),
+      })
+
+      return <div>{`status: ${mutation.status}`}</div>
+    }
+
+    const markup = renderToString(
+      <QueryClientProvider client={queryClient}>
+        <Page />
+      </QueryClientProvider>,
+    )
+
+    expect(markup).toContain('status: idle')
+
+    queryCache.clear()
+  })
+
+  it('useMutationState should return empty array', () => {
+    function Page() {
+      const mutationState = useMutationState()
+
+      return <div>{`mutationState: ${mutationState.length}`}</div>
+    }
+
+    const markup = renderToString(
+      <QueryClientProvider client={queryClient}>
+        <Page />
+      </QueryClientProvider>,
+    )
+
+    expect(markup).toContain('mutationState: 0')
 
     queryCache.clear()
   })