+++ packages/eslint-plugin-query/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tanstack/eslint-plugin-query
 
+## 5.91.4
+
+### Patch Changes
+
+- fix(eslint-plugin-query): declare typescript as optional peer dependency ([#10007](https://github.com/TanStack/query/pull/10007))
+
+## 5.91.3
+
+### Patch Changes
+
+- exhaustive-deps rule fixed for vue files ([#10011](https://github.com/TanStack/query/pull/10011))
+
 ## 5.91.2
 
 ### Patch Changes