+++ packages/svelte-query-persist-client/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @tanstack/svelte-query-persist-client
 
+## 6.0.22
+
+### Patch Changes
+
+- Updated dependencies [[`e505568`](https://github.com/TanStack/query/commit/e505568f4d51c8281d38e9687091094b7d32a405)]:
+  - @tanstack/query-persist-client-core@5.92.1
+
+## 6.0.21
+
+### Patch Changes
+
+- Updated dependencies [[`978fc52`](https://github.com/TanStack/query/commit/978fc52728a8b9eb33f0a82f4ddf42a95815bd7f)]:
+  - @tanstack/query-persist-client-core@5.92.0
+
+## 6.0.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.19
+  - @tanstack/svelte-query@6.0.18
+
+## 6.0.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.18
+  - @tanstack/svelte-query@6.0.17
+
+## 6.0.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.17
+  - @tanstack/svelte-query@6.0.16
+
+## 6.0.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.16
+  - @tanstack/svelte-query@6.0.15
+
 ## 6.0.16
 
 ### Patch Changes